# openapi.yaml
openapi: 3.0.3
info:
  title: SnapPwd Service API
  description: |
    A secure, ephemeral secret and file sharing service. Data is stored encrypted
    in Redis with automatic expiration and one-time access semantics - secrets and
    files are deleted immediately after retrieval.

    Secret creation also returns reveal receipt credentials. The service stores
    only a receipt token hash, never the raw receipt token. Receipt records live
    until the original secret expiration plus a 24 hour retention window, even if
    the secret is retrieved earlier.
  version: 1.0.0
  contact:
    name: OpenClaw

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /v1/secrets:
    post:
      summary: Create a secret
      description: |
        Store an encrypted secret with time-based expiration. The response adds
        `receiptId` and `receiptToken` alongside `secretId`; this is backward
        compatible for JSON clients that ignore unknown response fields.
      operationId: createSecret
      tags:
        - Secrets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SecretRequest'
      responses:
        '201':
          description: Secret created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecretResponse'
        '400':
          description: Invalid expiration time (must be between 60 and 2592000 seconds)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/secrets/{id}:
    get:
      summary: Retrieve a secret
      description: |
        Retrieve a secret. By default, the secret is deleted after retrieval (one-time access).
        Use `?peek=true` to view metadata without burning the secret.
      operationId: getSecret
      tags:
        - Secrets
      parameters:
        - name: id
          in: path
          required: true
          description: Secret ID (must start with "sps-")
          schema:
            type: string
            example: sps-7Yx8kN2mPqRs
        - name: peek
          in: query
          required: false
          description: |
            If true, returns metadata (creation time, TTL, custom metadata) without
            deleting the secret. The secret remains available for future access.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: |
            Secret retrieved successfully. Response format depends on the `peek` parameter:
            - `peek=false` (default): Returns the encrypted secret (secret is now deleted)
            - `peek=true`: Returns metadata only (secret is preserved)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/EncryptedSecretResponse'
                  - $ref: '#/components/schemas/SecretPeekResponse'
        '404':
          description: Secret not found or already accessed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/receipts/{receiptId}/status:
    post:
      summary: Get reveal receipt status
      description: |
        Authenticate with the receipt token returned during secret creation and
        return whether the secret has been accessed. The token is sent in the
        request body and is compared against the server-side token hash; raw
        receipt tokens are not stored.

        Receipt status values are:
        - `not_accessed`: the secret is still available and has not been retrieved.
        - `accessed`: the one-time destructive secret retrieval succeeded.
        - `expired_unavailable`: the secret was not accessed before expiration, or
          is otherwise unavailable during the receipt retention window.

        Invalid receipt IDs and invalid tokens return the same generic response
        so token failures do not reveal whether a receipt exists. All responses
        include `Cache-Control: no-store`.
      operationId: getReceiptStatus
      tags:
        - Receipts
      parameters:
        - name: receiptId
          in: path
          required: true
          description: Receipt ID returned by `POST /v1/secrets`
          schema:
            type: string
            pattern: ^spr-[A-Za-z0-9]+$
            example: spr-2Vj8kN2mPqRs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptStatusRequest'
      responses:
        '200':
          description: Receipt status returned for a valid receipt token
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptStatusResponse'
        '404':
          description: Invalid receipt credentials
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          headers:
            Cache-Control:
              schema:
                type: string
                example: no-store
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/files:
    post:
      summary: Create a file
      description: Store an encrypted file with metadata and time-based expiration
      operationId: createFile
      tags:
        - Files
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FileRequest'
      responses:
        '201':
          description: File created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileResponse'
        '400':
          description: Invalid request (expiration out of range or file too large)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/files/{id}:
    get:
      summary: Retrieve a file
      description: |
        Retrieve a file. By default, the file is deleted after retrieval (one-time access).
        Use `?peek=true` to view metadata without burning the file.
      operationId: getFile
      tags:
        - Files
      parameters:
        - name: id
          in: path
          required: true
          description: File ID (must start with "spf-")
          schema:
            type: string
            example: spf-9Zw3mK7nPqTs
        - name: peek
          in: query
          required: false
          description: |
            If true, returns metadata (creation time, TTL, file metadata) without
            deleting the file. The file remains available for future access.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: |
            File retrieved successfully. Response format depends on the `peek` parameter:
            - `peek=false` (default): Returns the full file with encrypted data (file is now deleted)
            - `peek=true`: Returns metadata only (file is preserved)
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/StoredFile'
                  - $ref: '#/components/schemas/FilePeekResponse'
        '404':
          description: File not found or already accessed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    SecretRequest:
      type: object
      required:
        - encryptedSecret
        - expiration
      properties:
        encryptedSecret:
          type: string
          description: The encrypted secret data
          example: "U2FsdGVkX1+vupppZksvRf5pq5g5XjFRlipGLw=="
        expiration:
          type: integer
          format: int64
          minimum: 60
          maximum: 2592000
          description: Time in seconds until the secret expires (1 minute to 30 days)
          example: 3600
        metadata:
          type: object
          nullable: true
          description: Optional custom metadata to store with the secret (viewable via peek)
          example: {"label": "API key for staging"}

    SecretResponse:
      type: object
      description: |
        Secret creation response. `receiptId` and `receiptToken` are additive
        fields; clients that only consume `secretId` remain compatible.
      required:
        - secretId
        - receiptId
        - receiptToken
      properties:
        secretId:
          type: string
          description: Unique identifier for retrieving the secret
          pattern: ^sps-[A-Za-z0-9]+$
          example: sps-7Yx8kN2mPqRs
        receiptId:
          type: string
          description: Identifier for querying reveal receipt status
          pattern: ^spr-[A-Za-z0-9]+$
          example: spr-2Vj8kN2mPqRs
        receiptToken:
          type: string
          description: |
            Bearer-style receipt token required by the receipt status endpoint.
            Returned once at creation time; only its SHA-256 hash is stored.
          example: spt-4Gm9kP2xQwRt

    EncryptedSecretResponse:
      type: object
      required:
        - encryptedSecret
      properties:
        encryptedSecret:
          type: string
          description: The retrieved encrypted secret
          example: "U2FsdGVkX1+vupppZksvRf5pq5g5XjFRlipGLw=="

    SecretPeekResponse:
      type: object
      required:
        - createdAt
        - ttlSeconds
      properties:
        createdAt:
          type: integer
          format: int64
          description: Unix timestamp when the secret was created (0 for legacy secrets)
          example: 1706900000
        ttlSeconds:
          type: integer
          format: int64
          description: Remaining time-to-live in seconds (-1 if no expiration, -2 if key doesn't exist)
          example: 298
        metadata:
          type: object
          nullable: true
          description: Custom metadata stored with the secret (null for legacy secrets)
          example: {"label": "API key for staging"}

    ReceiptStatusRequest:
      type: object
      required:
        - receiptToken
      properties:
        receiptToken:
          type: string
          description: Receipt token returned by `POST /v1/secrets`
          example: spt-4Gm9kP2xQwRt

    ReceiptStatusResponse:
      type: object
      required:
        - receiptId
        - status
        - secretExpiresAt
        - retentionUntil
      properties:
        receiptId:
          type: string
          description: Receipt identifier
          pattern: ^spr-[A-Za-z0-9]+$
          example: spr-2Vj8kN2mPqRs
        status:
          type: string
          enum:
            - not_accessed
            - accessed
            - expired_unavailable
          description: Current reveal receipt state
          example: accessed
        accessedAt:
          type: integer
          format: int64
          nullable: true
          description: Unix timestamp when destructive retrieval succeeded; omitted otherwise.
          example: 1706900100
        secretExpiresAt:
          type: integer
          format: int64
          description: Original Unix expiration time for the secret
          example: 1706903600
        retentionUntil:
          type: integer
          format: int64
          description: |
            Unix timestamp when the receipt record expires. This is the original
            secret expiration plus 24 hours.
          example: 1706990000

    FileMetadata:
      type: object
      required:
        - originalFilename
        - contentType
        - iv
      properties:
        originalFilename:
          type: string
          description: Original filename
          example: document.pdf
        contentType:
          type: string
          description: MIME type of the file
          example: application/pdf
        iv:
          type: string
          description: Initialization vector for encryption
          example: "abc123def456"

    FileRequest:
      type: object
      required:
        - metadata
        - encryptedData
        - expiration
      properties:
        metadata:
          $ref: '#/components/schemas/FileMetadata'
        encryptedData:
          type: string
          description: Base64-encoded encrypted file content
          example: "SGVsbG8gV29ybGQh..."
        expiration:
          type: integer
          format: int64
          minimum: 60
          maximum: 2592000
          description: Time in seconds until the file expires (1 minute to 30 days)
          example: 3600

    FileResponse:
      type: object
      required:
        - fileId
      properties:
        fileId:
          type: string
          description: Unique identifier for retrieving the file
          pattern: ^spf-[A-Za-z0-9]+$
          example: spf-9Zw3mK7nPqTs

    StoredFile:
      type: object
      required:
        - metadata
        - encryptedData
      properties:
        metadata:
          $ref: '#/components/schemas/FileMetadata'
        encryptedData:
          type: string
          description: Base64-encoded encrypted file content
          example: "SGVsbG8gV29ybGQh..."

    FilePeekResponse:
      type: object
      required:
        - createdAt
        - ttlSeconds
        - metadata
      properties:
        createdAt:
          type: integer
          format: int64
          description: Unix timestamp when the file was created (0 for legacy files)
          example: 1706900000
        ttlSeconds:
          type: integer
          format: int64
          description: Remaining time-to-live in seconds (-1 if no expiration, -2 if key doesn't exist)
          example: 298
        metadata:
          $ref: '#/components/schemas/FileMetadata'

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message describing what went wrong
          example: "Invalid expiration: must be between 60 and 2592000 seconds"

tags:
  - name: Secrets
    description: One-time secret sharing endpoints
  - name: Receipts
    description: Reveal receipt status endpoints
  - name: Files
    description: One-time file sharing endpoints